+++ src/components/CardForm.tsx
@@ -32,10 +32,10 @@ export interface Props extends AccessibilityProps {
   style?: StyleProp<ViewStyle>;
   autofocus?: boolean;
   testID?: string;
-
+  /** Applies a disabled state such that user input is not accepted. Defaults to false. */
+  disabled?: boolean;
   /** All styles except backgroundColor, cursorColor, borderColor, and borderRadius are Android only */
   cardStyle?: CardFormView.Styles;
-  // isUserInteractionEnabled?: boolean;
 
   // TODO: will make it public when iOS SDK allows for this
   // postalCodeEnabled?: boolean;
@@ -78,7 +78,6 @@ export const CardForm = forwardRef<CardFormView.Methods, Props>(
     {
       onFormComplete,
       cardStyle,
-      // isUserInteractionEnabled = true,
       // postalCodeEnabled = true,
       // onFocus,
       // onBlur,
@@ -181,7 +180,6 @@ export const CardForm = forwardRef<CardFormView.Methods, Props>(
           // disabledBackgroundColor: cardStyle?.disabledBackgroundColor,
           // type: cardStyle?.type,
         }}
-        // isUserInteractionEnabledValue={isUserInteractionEnabled}
         placeholders={{
           number: placeholders?.number,
           expiration: placeholders?.expiration,